requirements.txt
opencv-python==4.9.0.80
ultralytics==8.1.0
numpy==1.24.3
pillow==10.2.0



flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-engineio==4.8.0
python-socketio==5.10.0